Backup of dead end
This is a dead end, not only do we lose MSAA, but also depth surfaces
can not be shared directly between D3D9Ex and D3D11.

Maybe I can use a VirtualGL approach instead (and share the D3D9Ex to GL
inside CS:GO with NV_DX_INTEROP), but this has to be thought through
properly.