Lay foundation for fixing issue-861
I don't want to tamper agressively with mat_queue_mode anymore, it
causes too many problems.
Instead screen capturing and drawing will happen at the correct time on
the render thread, triggered by the engine thread, since we utilize the
render queue.
This means more narrow mod-support for features that depend on this, but
I think it's for the better.

TODO: Make campath drawer understand the queued drawing and remove
mat_queue_mode changes for that.