Recording props
This commit is more of an idea of what a feature of AGR could look like. It's useful for Garrys Mod users, but I won't be mad if we add them in a future update.

prop_physics are being used in many ways in Garrys Mod, they are a crucial part of the gameplay.
Since they can be used with weapons such as the gravity gun, I thought they could belong to projectiles.

Vehicles on the other hand, become what the player controls once entered, so they are kind of the "player".